style: apply pre-commit formatting to the whole snapshot tree
The gitcode pre-commit job runs clang-format, ruff format,
trailing-whitespace and end-of-file-fixer against the full PR file list.
This MR replaces the whole tree with the GitHub LLVM19 snapshot, so
thousands of historical files need formatting before those hooks pass.

Run the pre-commit formatters once over the tree (verified: all hooks
pass on the second run) and commit the result. Also bump the
check-added-large-files limit to 1000 KB for the three large generated
C++ sources (PTO.cpp 733KB, VMIToVPTO.cpp 579KB, PTOToEmitC.cpp 560KB).

+++ test/tilelang_st/npu/a5/src/st/smoke/testcase/trowargmin/gen_data.py
@@ -25,14 +25,22 @@ for case in CASES:
 
     if dtype in (np.int8, np.uint8, np.int16, np.uint16, np.int32, np.uint32):
         dtype_info = np.iinfo(dtype)
-        input1 = np.random.randint(dtype_info.min, dtype_info.max, size=shape).astype(dtype)
+        input1 = np.random.randint(dtype_info.min, dtype_info.max, size=shape).astype(
+            dtype
+        )
     else:
         dtype_info = np.finfo(dtype)
-        input1 = np.random.uniform(low=dtype_info.min, high=dtype_info.max, size=shape).astype(dtype)
+        input1 = np.random.uniform(
+            low=dtype_info.min, high=dtype_info.max, size=shape
+        ).astype(dtype)
 
     out_shape = (valid_shape[0], 1)
     golden = np.zeros(out_shape, dtype=dst_dtype)
-    golden[:, 0:1] = np.argmin(input1[:, :valid_shape[1]], axis=1, keepdims=True).astype(dst_dtype)
+    golden[:, 0:1] = np.argmin(
+        input1[:, : valid_shape[1]], axis=1, keepdims=True
+    ).astype(dst_dtype)
 
     save_case_data(case["name"], {"input1": input1, "golden": golden})
-    print(f"[INFO] gen_data: {case['name']} shape={shape} valid_shape={valid_shape} dtype={dtype.__name__}")
+    print(
+        f"[INFO] gen_data: {case['name']} shape={shape} valid_shape={valid_shape} dtype={dtype.__name__}"
+    )
