+++ test/tilelang_st/npu/a5/src/st/testcase/tmax/gen_data.py
@@ -26,7 +26,11 @@ for case in CASES:
 
     golden = np.zeros(shape, dtype=dtype)
     vr, vc = valid_shape
-    golden[:vr, :vc] = np.maximum(input1[:vr, :vc], input2[:vr, :vc]).astype(dtype, copy=False)
+    golden[:vr, :vc] = np.maximum(input1[:vr, :vc], input2[:vr, :vc]).astype(
+        dtype, copy=False
+    )
 
     save_case_data(case["name"], {"input1": input1, "input2": input2, "golden": golden})
-    print(f"[INFO] gen_data: {case['name']} shape={shape} valid_shape={valid_shape} dtype={dtype.__name__}")
+    print(
+        f"[INFO] gen_data: {case['name']} shape={shape} valid_shape={valid_shape} dtype={dtype.__name__}"
+    )
