+++ test/tilelang_st/npu/a5/src/st/testcase/tmins/gen_data.py
@@ -29,7 +29,11 @@ for case in CASES:
     golden = np.zeros(shape, dtype=dtype)
     vr, vc = valid_shape
     scalar_val = dtype(SCALAR)
-    golden[:vr, :vc] = np.minimum(input1[:vr, :vc], scalar_val).astype(dtype, copy=False)
+    golden[:vr, :vc] = np.minimum(input1[:vr, :vc], scalar_val).astype(
+        dtype, copy=False
+    )
 
     save_case_data(case["name"], {"input1": input1, "golden": golden})
-    print(f"[INFO] gen_data: {case['name']} shape={shape} valid_shape={valid_shape} dtype={dtype.__name__} scalar={SCALAR}")
+    print(
+        f"[INFO] gen_data: {case['name']} shape={shape} valid_shape={valid_shape} dtype={dtype.__name__} scalar={SCALAR}"
+    )
